requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
environs==9.5.0
frozenlist==1.3.0
idna==3.3
marshmallow==3.17.0
multidict==6.0.2
packaging==21.3
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
yarl==1.7.2
YooMoney==0.1.0
